Compact schema distribution on write is implemented.
hazelcast-enterprise-VERSION-tests.jar is downloaded from private-test-artifacts repo.

Enterprise tests are added to CLASSPATH

rc-bat corrected.
