+++ hazelcast/generated-sources/src/hazelcast/client/protocol/codec/codecs.h
@@ -3079,6 +3079,18 @@ ClientMessage HAZELCAST_API
 sql_fetch_encode(const sql::impl::query_id& query_id,
                  int32_t cursor_buffer_size);
 
+/**
+ * Replicates schema on cluster
+ */
+ClientMessage HAZELCAST_API
+send_schema_request_encode(const serialization::pimpl::schema& sch);
+
+/**
+ * Decodes response of send schema request
+ */
+std::vector<boost::uuids::uuid> HAZELCAST_API
+send_schema_response_decode(ClientMessage&);
+
 } // namespace codec
 } // namespace protocol
 } // namespace client
