+++ hazelcast/include/hazelcast/client/protocol/ClientMessage.h
@@ -27,6 +27,8 @@
 #include <string>
 #include <unordered_map>
 #include <vector>
+#include <iterator>
+#include <algorithm>
 
 #include <boost/endian/arithmetic.hpp>
 #include <boost/endian/conversion.hpp>
@@ -47,6 +49,8 @@
 #include "hazelcast/client/sql/impl/sql_error.h"
 #include "hazelcast/client/sql/sql_column_type.h"
 #include "hazelcast/client/protocol/codec/builtin/custom_type_factory.h"
+#include "hazelcast/client/serialization/pimpl/compact/schema.h"
+#include "hazelcast/client/serialization/pimpl/compact/field_descriptor.h"
 
 namespace hazelcast {
 namespace util {
@@ -1207,6 +1211,10 @@ public:
         header->flags = is_final ? IS_FINAL_FLAG : DEFAULT_FLAGS;
         std::memcpy(
           fp + SIZE_OF_FRAME_LENGTH_AND_FLAGS, &bytes[0], bytes.size());
+
+        copy(begin(value.schemas_will_be_replicated()),
+             end(value.schemas_will_be_replicated()),
+             back_inserter(schemas_will_be_replicated_));
     }
 
     inline void set(const serialization::pimpl::data* value,
@@ -1284,6 +1292,44 @@ public:
         add_end_frame(is_final);
     }
 
+    void set(const serialization::pimpl::field_descriptor& descriptor,
+             const std::string& field_name,
+             bool is_final = false)
+    {
+        add_begin_frame();
+
+        set(frame_header_type{ SIZE_OF_FRAME_LENGTH_AND_FLAGS + INT32_SIZE,
+                               DEFAULT_FLAGS });
+        set(int32_t(descriptor.kind));
+        set(field_name);
+
+        add_end_frame(is_final);
+    }
+
+    void set(const serialization::pimpl::schema& s, bool is_final = false)
+    {
+        add_begin_frame();
+
+        set(s.type_name());
+
+        { // Fields list
+            add_begin_frame();
+
+            for (const auto& p : s.fields()) {
+                const std::string& field_name{ p.first };
+                const serialization::pimpl::field_descriptor& descriptor{
+                    p.second
+                };
+
+                set(descriptor, field_name, false);
+            }
+
+            add_end_frame(false);
+        }
+
+        add_end_frame(is_final);
+    }
+
     //----- Setter methods end ---------------------
 
     //----- utility methods -------------------
@@ -1342,6 +1388,8 @@ public:
     }
 
     void fast_forward_to_end_frame();
+    const std::vector<serialization::pimpl::schema>&
+    schemas_will_be_replicated() const;
 
     static const frame_header_type& null_frame();
     static const frame_header_type& begin_frame();
@@ -1446,6 +1494,7 @@ private:
     std::vector<std::vector<byte>> data_buffer_;
     size_t buffer_index_{ 0 };
     size_t offset_{ 0 };
+    std::vector<serialization::pimpl::schema> schemas_will_be_replicated_;
 };
 
 template<>
