+++ hazelcast/include/hazelcast/client/spi/impl/ClientInvocation.h
@@ -32,6 +32,8 @@
 #include "hazelcast/client/spi/EventHandler.h"
 #include "hazelcast/client/protocol/ClientMessage.h"
 
+#include "hazelcast/client/serialization/pimpl/compact/default_schema_service.h"
+
 #if defined(WIN32) || defined(_WIN32) || defined(WIN64) || defined(_WIN64)
 #pragma warning(push)
 #pragma warning(disable : 4251) // for dll export
@@ -152,10 +154,14 @@ private:
     static constexpr int MAX_FAST_INVOCATION_COUNT = 5;
     static constexpr int UNASSIGNED_PARTITION = -1;
 
+    boost::future<boost::csbl::vector<boost::future<void>>> replicate_schemas(
+      const std::vector<serialization::pimpl::schema>& schemas);
+
     logger& logger_;
     lifecycle_service& lifecycle_service_;
     ClientInvocationServiceImpl& invocation_service_;
     std::shared_ptr<ClientExecutionServiceImpl> execution_service_;
+    serialization::pimpl::default_schema_service& schema_service_;
     boost::atomic_shared_ptr<std::shared_ptr<protocol::ClientMessage>>
       client_message_;
     std::shared_ptr<sequence::CallIdSequence> call_id_sequence_;
