+++ hazelcast/src/hazelcast/client/spi.cpp
@@ -325,6 +325,12 @@ ClientContext::get_proxy_session_manager()
     return hazelcast_client_.proxy_session_manager_;
 }
 
+serialization::pimpl::default_schema_service&
+ClientContext::get_schema_service()
+{
+    return hazelcast_client_.schema_service_;
+}
+
 lifecycle_service::lifecycle_service(
   ClientContext& client_context,
   const std::vector<lifecycle_listener>& listeners)
@@ -1357,6 +1363,7 @@ ClientInvocation::ClientInvocation(
   , invocation_service_(client_context.get_invocation_service())
   , execution_service_(
       client_context.get_client_execution_service().shared_from_this())
+  , schema_service_(client_context.get_schema_service())
   , call_id_sequence_(client_context.get_call_id_sequence())
   , uuid_(uuid)
   , partition_id_(partition)
@@ -1381,20 +1388,44 @@ boost::future<protocol::ClientMessage>
 ClientInvocation::invoke()
 {
     assert(client_message_.load());
-    // for back pressure
-    call_id_sequence_->next();
-    invoke_on_selection();
-    if (!lifecycle_service_.is_running()) {
+
+    auto actual_work = [this]() {
+        // for back pressure
+        call_id_sequence_->next();
+        invoke_on_selection();
+        if (!lifecycle_service_.is_running()) {
+            return invocation_promise_.get_future().then(
+              [](boost::future<protocol::ClientMessage> f) { return f.get(); });
+        }
+        auto id_seq = call_id_sequence_;
         return invocation_promise_.get_future().then(
-          [](boost::future<protocol::ClientMessage> f) { return f.get(); });
+          execution_service_->get_user_executor(),
+          [=](boost::future<protocol::ClientMessage> f) {
+              id_seq->complete();
+              return f.get();
+          });
+    };
+
+    const auto& schemas =
+      (*(client_message_.load()))->schemas_will_be_replicated();
+
+    if (!schemas.empty()) {
+        auto self = shared_from_this();
+
+        return replicate_schemas(schemas)
+          .then(
+            boost::launch::sync,
+            [this, actual_work, self](
+              boost::future<boost::csbl::vector<boost::future<void>>> replications) {
+                for (auto& replication : replications.get())
+                    replication.get();
+
+                return actual_work();
+            })
+          .unwrap();
     }
-    auto id_seq = call_id_sequence_;
-    return invocation_promise_.get_future().then(
-      execution_service_->get_user_executor(),
-      [=](boost::future<protocol::ClientMessage> f) {
-          id_seq->complete();
-          return f.get();
-      });
+
+    return actual_work();
 }
 
 boost::future<protocol::ClientMessage>
@@ -1402,20 +1433,41 @@ ClientInvocation::invoke_urgent()
 {
     assert(client_message_.load());
     urgent_ = true;
+
     // for back pressure
     call_id_sequence_->force_next();
     invoke_on_selection();
     if (!lifecycle_service_.is_running()) {
         return invocation_promise_.get_future().then(
-          [](boost::future<protocol::ClientMessage> f) { return f.get(); });
+            [](boost::future<protocol::ClientMessage> f) { return f.get(); });
     }
     auto id_seq = call_id_sequence_;
     return invocation_promise_.get_future().then(
-      execution_service_->get_user_executor(),
-      [=](boost::future<protocol::ClientMessage> f) {
-          id_seq->complete();
-          return f.get();
-      });
+        execution_service_->get_user_executor(),
+        [=](boost::future<protocol::ClientMessage> f) {
+            id_seq->complete();
+            return f.get();
+        });
+}
+
+boost::future<boost::csbl::vector<boost::future<void>>>
+ClientInvocation::replicate_schemas(
+  const std::vector<serialization::pimpl::schema>& schemas)
+{
+    std::vector<boost::future<void>> replications;
+
+    replications.reserve(schemas.size());
+
+    transform(begin(schemas),
+              end(schemas),
+              back_inserter(replications),
+              [this](const serialization::pimpl::schema& s) {
+                  return schema_service_.replicate_schema(s);
+              });
+
+    auto self = shared_from_this();
+
+    return boost::when_all(begin(replications), end(replications));
 }
 
 void
