+++ hazelcast/test/resources/compact.xml
@@ -1,0 +1,5 @@
+<hazelcast xmlns="http://www.hazelcast.com/schema/config">
+
+    <cluster-name>compact-dev</cluster-name>
+
+</hazelcast>
