+++ hazelcast/test/src/ClientTest.cpp
@@ -29,6 +29,7 @@
 
 #include "HazelcastServerFactory.h"
 #include "remote_controller_client.h"
+#include "TestHelperFunctions.h"
 
 namespace hazelcast {
 namespace client {
@@ -105,16 +106,6 @@ std::string
 ClientTest::random_map_name()
 {
     return "test_" + boost::replace_all_copy(random_string(), "-", "_");
-}
-
-std::string
-ClientTest::random_string()
-{
-    // performance is not important, hence we can use random_device for the
-    // tests
-    std::random_device rand{};
-    return boost::uuids::to_string(
-      boost::uuids::basic_random_generator<std::random_device>{ rand }());
 }
 
 boost::uuids::uuid
