+++ hazelcast/test/src/HazelcastTests2.cpp
@@ -834,11 +834,13 @@ public:
 TEST_F(PortableVersionTest, test_nestedPortable_versionedSerializer)
 {
     serialization_config serializationConfig;
-    serialization::pimpl::SerializationService ss1(serializationConfig);
+    serialization::pimpl::SerializationService ss1(serializationConfig,
+                                                   null_schema_service());
 
     serialization_config serializationConfig2;
     serializationConfig2.set_portable_version(6);
-    serialization::pimpl::SerializationService ss2(serializationConfig2);
+    serialization::pimpl::SerializationService ss2(serializationConfig2,
+                                                   null_schema_service());
 
     // make sure ss2 cached class definition of Child
     {
@@ -932,7 +934,7 @@ TEST_F(PartitionAwareTest, testSimplePartitionAwareObjectSerialisation)
 {
     serialization_config serializationConfig;
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     SimplePartitionAwareObject obj;
     serialization::pimpl::data data =
@@ -950,7 +952,7 @@ TEST_F(PartitionAwareTest, testNonPartitionAwareObjectSerialisation)
 {
     serialization_config serializationConfig;
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     int obj = 7;
     serialization::pimpl::data data = serializationService.to_data<int>(&obj);
@@ -989,7 +991,7 @@ class JsonValueSerializationTest : public ::testing::Test
 {
 public:
     JsonValueSerializationTest()
-      : serialization_service_(config_)
+      : serialization_service_(config_, null_schema_service())
     {}
 
 protected:
@@ -1082,7 +1084,7 @@ TEST_F(ClientSerializationTest, testCustomSerialization)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     TestCustomXSerializable a{ 131321 };
     serialization::pimpl::data data =
@@ -1104,7 +1106,7 @@ TEST_F(ClientSerializationTest, testRawData)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
     char charA[] = "test chars";
     std::vector<char> chars(charA, charA + 10);
     std::vector<byte> bytes;
@@ -1125,7 +1127,7 @@ TEST_F(ClientSerializationTest, testIdentifiedDataSerializable)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
     serialization::pimpl::data data;
     TestDataSerializable np{ 4, 'k' };
     data = serializationService.to_data<TestDataSerializable>(&np);
@@ -1146,7 +1148,7 @@ TEST_F(ClientSerializationTest, testRawDataWithoutRegistering)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
     char charA[] = "test chars";
     std::vector<char> chars(charA, charA + 10);
     std::vector<byte> bytes;
@@ -1167,7 +1169,7 @@ TEST_F(ClientSerializationTest, testInvalidWrite)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
     TestInvalidWritePortable p{ 2131, 123, "q4edfd" };
     ASSERT_THROW(serializationService.to_data<TestInvalidWritePortable>(&p),
                  exception::hazelcast_serialization);
@@ -1178,7 +1180,7 @@ TEST_F(ClientSerializationTest, testInvalidRead)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
     TestInvalidReadPortable p{ 2131, 123, "q4edfd" };
     serialization::pimpl::data data =
       serializationService.to_data<TestInvalidReadPortable>(&p);
@@ -1191,12 +1193,12 @@ TEST_F(ClientSerializationTest, testDifferentVersions)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     serialization_config serializationConfig2;
     serializationConfig2.set_portable_version(2);
     serialization::pimpl::SerializationService serializationService2(
-      serializationConfig2);
+      serializationConfig2, null_schema_service());
 
     serialization::pimpl::data data =
       serializationService.to_data<TestNamedPortable>(
@@ -1223,7 +1225,7 @@ TEST_F(ClientSerializationTest, testBasicFunctionality)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
     serialization::pimpl::data data;
 
     int x = 3;
@@ -1279,7 +1281,8 @@ TEST_F(ClientSerializationTest, testStringLiterals)
 {
     auto literal = R"delimeter(My example string literal)delimeter";
     serialization_config config;
-    serialization::pimpl::SerializationService serializationService(config);
+    serialization::pimpl::SerializationService serializationService(
+      config, null_schema_service());
     auto data = serializationService.to_data(literal);
     auto obj = serializationService.to_object<decltype(literal)>(data);
     ASSERT_TRUE(obj);
@@ -1291,7 +1294,7 @@ TEST_F(ClientSerializationTest, testBasicFunctionalityWithLargeData)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
     serialization::pimpl::data data;
 
     std::vector<byte> bb(LARGE_ARRAY_SIZE);
@@ -1327,12 +1330,12 @@ TEST_F(ClientSerializationTest, testBasicFunctionalityWithDifferentVersions)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     serialization_config serializationConfig2;
     serializationConfig2.set_portable_version(2);
     serialization::pimpl::SerializationService serializationService2(
-      serializationConfig2);
+      serializationConfig2, null_schema_service());
     serialization::pimpl::data data;
 
     int32_t x = 3;
@@ -1389,7 +1392,7 @@ TEST_F(ClientSerializationTest, testDataHash)
 {
     serialization_config serializationConfig;
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
     std::string serializable = "key1";
     serialization::pimpl::data data =
       serializationService.to_data<std::string>(&serializable);
@@ -1402,7 +1405,7 @@ TEST_F(ClientSerializationTest, testPrimitives)
 {
     serialization_config serializationConfig;
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
     byte by = 2;
     bool boolean = true;
     char c = 'c';
@@ -1431,7 +1434,7 @@ TEST_F(ClientSerializationTest, testPrimitiveArrays)
 {
     serialization_config serializationConfig;
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     char charArray[] = { 'c', 'h', 'a', 'r' };
     std::vector<char> cc(charArray, charArray + 4);
@@ -1489,7 +1492,8 @@ TEST_F(ClientSerializationTest, testPrimitiveArrays)
 TEST_F(ClientSerializationTest, testWriteObjectWithPortable)
 {
     serialization_config serializationConfig;
-    serialization::pimpl::SerializationService ss(serializationConfig);
+    serialization::pimpl::SerializationService ss(serializationConfig,
+                                                  null_schema_service());
 
     ObjectCarryingPortable<TestNamedPortable> objectCarryingPortable{
         TestNamedPortable{ "name", 2 }
@@ -1504,7 +1508,8 @@ TEST_F(ClientSerializationTest, testWriteObjectWithPortable)
 TEST_F(ClientSerializationTest, testWriteObjectWithIdentifiedDataSerializable)
 {
     serialization_config serializationConfig;
-    serialization::pimpl::SerializationService ss(serializationConfig);
+    serialization::pimpl::SerializationService ss(serializationConfig,
+                                                  null_schema_service());
 
     ObjectCarryingPortable<TestDataSerializable> objectCarryingPortable{
         TestDataSerializable{ 2, 'c' }
@@ -1519,7 +1524,8 @@ TEST_F(ClientSerializationTest, testWriteObjectWithIdentifiedDataSerializable)
 TEST_F(ClientSerializationTest, testWriteObjectWithCustomXSerializable)
 {
     serialization_config serializationConfig;
-    serialization::pimpl::SerializationService ss(serializationConfig);
+    serialization::pimpl::SerializationService ss(serializationConfig,
+                                                  null_schema_service());
     ObjectCarryingPortable<TestCustomXSerializable> objectCarryingPortable{
         TestCustomXSerializable{ 131321 }
     };
@@ -1534,7 +1540,8 @@ TEST_F(ClientSerializationTest, testWriteObjectWithCustomXSerializable)
 TEST_F(ClientSerializationTest, testWriteObjectWithCustomPersonSerializable)
 {
     serialization_config serializationConfig;
-    serialization::pimpl::SerializationService ss(serializationConfig);
+    serialization::pimpl::SerializationService ss(serializationConfig,
+                                                  null_schema_service());
 
     ObjectCarryingPortable<TestCustomPerson> objectCarryingPortable{
         TestCustomPerson{ "TestCustomPerson" }
@@ -1550,7 +1557,8 @@ TEST_F(ClientSerializationTest, testNullData)
 {
     serialization::pimpl::data data;
     serialization_config serializationConfig;
-    serialization::pimpl::SerializationService ss(serializationConfig);
+    serialization::pimpl::SerializationService ss(serializationConfig,
+                                                  null_schema_service());
     auto ptr = ss.to_object<int32_t>(data);
     ASSERT_FALSE(ptr.has_value());
 }
@@ -1560,7 +1568,7 @@ TEST_F(ClientSerializationTest, testMorphingPortableV1ToV2Conversion)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     TestNamedPortable p{ "portable-v1", 123 };
     serialization::pimpl::data data =
@@ -1578,7 +1586,7 @@ TEST_F(ClientSerializationTest, testMorphingPortableV2ToV1Conversion)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     TestNamedPortableV2 p2{ "portable-v2", 123, 9999 };
     serialization::pimpl::data data =
@@ -1595,7 +1603,7 @@ TEST_F(ClientSerializationTest, testMorphingPortableV1ToV3Conversion)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     TestNamedPortable p{ "portable-v1", 123 };
     serialization::pimpl::data data =
@@ -1611,12 +1619,12 @@ TEST_F(ClientSerializationTest,
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     serialization_config serializationConfig2;
     serializationConfig.set_portable_version(5);
     serialization::pimpl::SerializationService serializationService2(
-      serializationConfig2);
+      serializationConfig2, null_schema_service());
 
     TestNamedPortableV2 p2{ "portable-v2", 123, 7 };
     serialization::pimpl::data data2 =
@@ -1633,7 +1641,7 @@ TEST_F(ClientSerializationTest, object_data_input_output)
     serialization_config serializationConfig;
     serializationConfig.set_portable_version(1);
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     serialization::object_data_output out(boost::endian::order::big);
 
@@ -1749,7 +1757,7 @@ TEST_F(ClientSerializationTest, testExtendedAscii)
 
     serialization_config serializationConfig;
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     serialization::pimpl::data data =
       serializationService.to_data<std::string>(&utfStr);
@@ -1764,7 +1772,7 @@ TEST_F(ClientSerializationTest, testGlobalSerializer)
     serializationConfig.set_global_serializer(
       std::make_shared<DummyGlobalSerializer>());
     serialization::pimpl::SerializationService serializationService(
-      serializationConfig);
+      serializationConfig, null_schema_service());
 
     NonSerializableObject obj{ "My class with no serializer" };
 
@@ -2151,7 +2159,7 @@ public:
     {
         ss_ = std::unique_ptr<serialization::pimpl::SerializationService>(
           new serialization::pimpl::SerializationService(
-            serialization_config_));
+            serialization_config_, null_schema_service()));
     }
 
 protected:
