+++ hazelcast/test/src/compact/serialization/empty_main_dto.h
@@ -1,0 +1,57 @@
+/*
+ * Copyright (c) 2008-2022, Hazelcast, Inc. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+#include "hazelcast/client/serialization/serialization.h"
+
+namespace hazelcast {
+namespace client {
+namespace test {
+namespace compact {
+
+/**
+ * This class is to simulate versioning.
+ * We will provide this struct with serializer returning type name of the
+ * original main dto. This way we can use the serialized data of this class
+ * to test to_object of the original main_dto.
+ */
+struct empty_main_dto
+{};
+
+}
+}
+
+namespace serialization {
+
+template<>
+struct hz_serializer<compact::test::empty_main_dto> : public compact_serializer
+{
+    static void write(const compact::test::empty_main_dto& object,
+                      compact_writer& writer)
+    {}
+
+    static compact::test::empty_main_dto read(compact_reader& reader)
+    {
+        return compact::test::empty_main_dto{};
+    }
+
+    static std::string type_name() { return "main"; }
+};
+
+}
+}
+}
